V.220724

+++ yatube/posts/views.py
@@ -3,15 +3,15 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db.models import Q
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse_lazy
-from django.views.decorators.cache import cache_page
+# from django.views.decorators.cache import cache_page
 from django.views.generic.edit import DeleteView
 
 from .forms import CommentForm, PostForm
-from .models import Follow, Group, Post, User
+from .models import Comment, Follow, Group, Post, User
 from .utils import paginator
 
 
-@cache_page(20, key_prefix='index_page')
+# @cache_page(20, key_prefix='index_page')
 def index(request):
     search_query = request.GET.get('search', '')
     if search_query:
@@ -27,7 +27,8 @@ def index(request):
     else:
         post_list = Post.objects.select_related('group')
     page_obj = paginator(post_list, request)
-    context = {'page_obj': page_obj, }
+    comments = Comment.objects.select_related('post')
+    context = {'page_obj': page_obj, 'comments': comments, }
     return render(request, 'posts/index.html', context)
 
 
@@ -35,9 +36,11 @@ def group_posts(request, slug):
     group = get_object_or_404(Group, slug=slug)
     post_list = group.posts.select_related()
     page_obj = paginator(post_list, request)
+    comments = Comment.objects.select_related('post')
     context = {
         'group': group,
         'page_obj': page_obj,
+        'comments': comments,
     }
     return render(request, 'posts/group_list.html', context)
 
@@ -47,17 +50,13 @@ def profile(request, username):
     author = get_object_or_404(User, username=username)
     post_list = author.posts.all()
     page_obj = paginator(post_list, request)
-    if request.user.is_authenticated and Follow.objects.filter(
-        user=user,
-        author=author
-    ).exists():
-        following = True
-    else:
-        following = False
+    following = user.is_authenticated and author.following.exists()
+    comments = Comment.objects.select_related('post')
     context = {
         'author': author,
         'page_obj': page_obj,
         'following': following,
+        'comments': comments,
     }
     return render(request, 'posts/profile.html', context)
 
@@ -126,9 +125,11 @@ def follow_index(request):
     authors = user.follower.values_list('author', flat=True)
     post_list = Post.objects.filter(author__id__in=authors)
     page_obj = paginator(post_list, request)
+    comments = Comment.objects.select_related('post')
     context = {
         'page_obj': page_obj,
         'user': user,
+        'comments': comments,
     }
     return render(request, 'posts/follow_index.html', context)
 
@@ -138,7 +139,7 @@ def profile_follow(request, username):
     """Функция подписки на автора."""
     author = User.objects.get(username=username)
     user = request.user
-    if user.id != author.id and user.id not in author.following.values_list(
+    if user.id != author.id and user not in author.following.values_list(
         'user',
         flat=True
     ):
@@ -151,8 +152,9 @@ def profile_unfollow(request, username):
     """Функция отмены подписки на автора."""
     author = User.objects.get(username=username)
     user = request.user
-    follow_relationship = Follow.objects.get(author=author, user=user)
-    follow_relationship.delete()
+    if Follow.objects.filter(user=user, author=author).exists():
+        follow_relationship = Follow.objects.get(author=author, user=user)
+        follow_relationship.delete()
     return redirect('posts:profile', username=author)
